Remove authentication, replace it with a simple LDAP-based allowlist.
Authentication is handled on the LB level, making sure that all requests
are made by users with a valid Mozilla LDAP account. However, for writing
state data, we want to ensure that only the WebCompat team has access,
so this implements a basic allow-list based on the LDAP email address.